% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfactors.R
\name{nfactors}
\alias{nfactors}
\title{Number of factors to be estimated in the RRmix model}
\usage{
nfactors(expr, var.exp = 0.8, plot = TRUE)
}
\arguments{
\item{expr}{An \code{n} by \code{G} matrix of \code{G} gene expression variables,
or compound abundances (standardized), on \code{n} samples.}

\item{var.exp}{The minimum cumulative proportion of the variance required to be
explained by the determined number of factors (\code{q.in}).  Float on (0,1)
(default = 0.8)}

\item{plot}{boolean.  Produce a scree plot of the variance explained by each
factor (default = \code{TRUE}).}
}
\value{
The number of factors that cumulatively explain >= \code{var.exp}
}
\description{
The number of factors to be used in the RRmix model such that
\code{var.exp} of the variance in the data is explained.
}
\details{
This function performs Principal Component Analysis (PCA) on the covariance
matrix for the raw data (\code{expr}).  \code{var.exp} sets the criteria for the
minimum variance explained by the determined number of factors (\code{q.in}) for the
RRmix model. \code{plot = TRUE}  produces a scree plot of the variance explained by
each factor.
}
\examples{
expr <- log(operators[, 1:12])  # Log-Transformed, First Two Operators
expr <- t(expr)                 # Transpose Matrix for n x G
q    <- nfactors(expr); q       # Number of Factors for 80\% Variance

}
\references{
Gao, C., Tignor, N. L., Salit, J., Strulovici-Barel, Y., Hackett, N. R., Crystal, R. G., and
Mezey, J. G. (2014). HEFT: eQTL analysis of many thousands of expressed genes while simultaneously
controlling for hidden factors. Bioinformatics, 30(3):369-376.
}

