% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\docType{data}
\name{operators}
\alias{operators}
\title{Operator Batch Effect Data Frame}
\format{A data frame with abundances of 265 filtered metabolites on 24 samples}
\source{
The Locasale Research Group - Department of Pharmacology & Cancer Biology, Duke University - Durham, North Carolina
}
\usage{
operators
}
\description{
This dataset contains the relative abundances of 265 metabolites across a total of 24 samples.
The samples were collected by four different operators, namely A, X, D, and Z.
Each operator performed a metabolomics experiment on 6 samples: three untreated (control) replicates, and three treated replicates.
Both operators A and X were given samples treated with the drug 2-deoxy-glucose (2DG).
Metabolites are in rows (265 metabolites), and samples are in columns (24 samples).
}
\examples{
dim(operators)      # [1] 265  24
summary(operators)  # Quantiles for each metabolite
}
\references{
NEED
}

