% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRmix.R
\docType{package}
\name{RRmix}
\alias{RRmix}
\alias{RRmix-package}
\title{RRmix: Model-Based Classification with Simultaneous Adjustment for Unwanted Variation}
\description{
NEED
}
\details{
NEED

The number of hidden factors \code{q} is determined using the function \code{nfactors},
which is based on the method described in Gao et al. (2014). Principal Component Analysis
(PCA) is performed on the correlation matrix of the raw data, and the number of factors is
chosen based on how many principal components explain \code{var.exp} of the total variation,
or through visual examination of the scree plot (\code{plot = TRUE}). Within a reasonable range
of values, \code{RRmix} is robust to the value of \code{q} (Wan 2015).
}
\author{
Muting Wan, Martin T. Wells, James G. Booth, Stephen Salerno
}
\references{
Gao, C., Tignor, N. L., Salit, J., Strulovici-Barel, Y., Hackett, N. R., Crystal, R. G., and
Mezey, J. G. (2014). HEFT: eQTL analysis of many thousands of expressed genes while simultaneously
controlling for hidden factors. Bioinformatics, 30(3):369-376.

M. Wan, "Model-based classification with applications to high-dimensional data in bioinformatics,"
Ph.D. dissertation, Cornell University, 2015.
}

