% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRmix.R
\name{runRRmix}
\alias{runRRmix}
\title{Emperical Bayes ECM Implementation of RRmix Algorithm}
\usage{
runRRmix(Y.in, SNP.in, Xc.in = matrix(nrow = 0, ncol = 0),
  betac.0 = matrix(nrow = 0, ncol = 0), sig20.0 = 1, sig21.0 = 0.1,
  p.0 = 0.05, er_tol.in = 10^(-3), q.in = 1)
}
\arguments{
\item{Y.in}{An \code{n} by \code{G} matrix of \code{G} gene expression variables, or compound abundances (standardized), on \code{n} samples.}

\item{SNP.in}{A \code{G} by \code{1} indicator vector for treatment condition or minor allele presence in a single SNP (standardized).}

\item{Xc.in}{\code{n} by \code{r} matrix of \code{r} known covariates (standardized, default = \code{matrix(nrow=0,ncol=0)}).}

\item{betac.0}{\code{n} by \code{r} matrix of initial values for EM estimated covariate term coefficients (default = \code{matrix(nrow=0,ncol=0)}).}

\item{sig20.0}{Initial value for EM estimated first variance component for \eqn{\beta_g | b_g} (default = 1.0).}

\item{sig21.0}{Initial value for EM estimated second variance component for \eqn{\beta_g | b_g} (default = 0.1).}

\item{p.0}{Initial value for EM estimated proportion of differentially abundant compounds (default = 0.05).}

\item{er_tol.in}{Convergence criterion for EM algorithm (default = 0.001).}

\item{q.in}{The number of factors to be estimated (default = 1).}
}
\value{
A list object containing the following named attributes:
\item{er.all}{Vector of error/tolerances at each iteration of the EM algorithm until convergence.}
\item{lc}{Vector of likelihood values at each iteration of the EM algorithm until convergence.}
\item{mu}{Vector of EM estimated sample-level means over genes}
\item{sig20}{EM estimated first variance component for \eqn{\beta_g | b_g}.}
\item{sig21}{EM estimated second variance component for \eqn{\beta_g | b_g}.}
\item{Lam}{Estimated \eqn{n\times q} loading matrix \eqn{(\Lambda)} for the factor analysis component of the model.}
\item{sig2_g}{Estimated compound-specific error variance.}
\item{b_g}{Vector of the posterior probabilities of differential abundance for each of the \code{G} compounds.}
\item{p}{EM estimated proportion of differentially abundant compounds.}
\item{psi}{EM estimated main effect of differential abundance.}
\item{beta_g.1}{Estimated \eqn{2\times G} coefficient matrix for the primary effect of treatment condition.}
\item{beta_g.0}{Initial \eqn{2\times G} coefficient matrix for the primary effect of treatment condition.}
\item{F_g.1}{Estimated \eqn{q\times G} factor matrix for the factor analysis component of the model.}
\item{F_g.0}{Initial \eqn{q\times G} factor matrix for the factor analysis component of the model.}
\item{assoc.coefs}{Vector of \code{G} associated coefficients for each of the \code{G} compounds.}
\item{sig20.0}{Initial value for EM estimated first variance component for \eqn{\beta_g | b_g}.}
\item{sig21.0}{Initial value for EM estimated second variance component for \eqn{\beta_g | b_g}.}
\item{p.0}{Initial value for EM estimated proportion of differentially abundant compounds.}
\item{psi.0}{Initial value for EM estimated main effect of differential abundance.}
}
\description{
This function provides a variant on the Expectation-Maximization (EM) Algorithm for the estimation
of the \code{RRmix} hierarchical mixture model parameters. Empirical Bayes inference is implemented via an EM
algorithm with classification based on the posterior expectation of the latent indicator variables. This model-based
classification method with simultaneous adjustment for unwanted variation is adapted to solve association detection
problems for high-dimensional data in the presence of unwanted variation.
}
\examples{
expr    <- log(operators[, 1:12])                      # Use log of First Two Operators' Abundances
expr    <- t(expr)                                     # Transpose Matrix for n x G
trmt    <- c(rep(0,3), rep(1,3), rep(0,3), rep(1,3))   # SNP Vector
results <- runRRmix(Y.in=expr, SNP.in=trmt, q.in=1)    # RRmix Results

names(results)                                         # Display Result Components

}
\references{
M. Wan, "Model-based classification with applications to high-dimensional data in bioinformatics,"
Ph.D. dissertation, Cornell University, 2015.
}

